feat: create a domain module that stores all entities of the project
