+++ domain/src/main/java/com/yourcity/domain/entity/Address.java
@@ -1,4 +1,4 @@
-package com.yourcity.yourcity.model.entity;
+package com.yourcity.domain.entity;
 
 import jakarta.persistence.*;
 import lombok.*;
