+++ domain/src/main/java/com/yourcity/domain/entity/Country.java
@@ -1,0 +1,29 @@
+package com.yourcity.domain.entity;
+
+import jakarta.persistence.Entity;
+import jakarta.persistence.GeneratedValue;
+import jakarta.persistence.Id;
+import lombok.AllArgsConstructor;
+import lombok.Builder;
+import lombok.Data;
+import lombok.NoArgsConstructor;
+import lombok.experimental.FieldDefaults;
+
+import static jakarta.persistence.GenerationType.IDENTITY;
+import static lombok.AccessLevel.PRIVATE;
+
+@Data
+@Builder
+@NoArgsConstructor
+@AllArgsConstructor
+@FieldDefaults(level = PRIVATE)
+@Entity
+public class Country {
+
+    @Id
+    @GeneratedValue(strategy = IDENTITY)
+    Long id;
+
+    String name;
+    String description;
+}
