+++ domain/src/main/java/com/yourcity/domain/entity/Street.java
@@ -1,0 +1,33 @@
+package com.yourcity.domain.entity;
+
+import com.yourcity.domain.enums.Type;
+import jakarta.persistence.Entity;
+import jakarta.persistence.Enumerated;
+import jakarta.persistence.GeneratedValue;
+import jakarta.persistence.Id;
+import lombok.AllArgsConstructor;
+import lombok.Builder;
+import lombok.Data;
+import lombok.NoArgsConstructor;
+import lombok.experimental.FieldDefaults;
+
+import static jakarta.persistence.EnumType.STRING;
+import static jakarta.persistence.GenerationType.IDENTITY;
+import static lombok.AccessLevel.PRIVATE;
+
+@Data
+@Builder
+@NoArgsConstructor
+@AllArgsConstructor
+@FieldDefaults(level = PRIVATE)
+@Entity
+public class Street {
+
+    @Id
+    @GeneratedValue(strategy = IDENTITY)
+    Long id;
+
+    @Enumerated(STRING)
+    Type type;
+    String name;
+}
