+++ domain/src/main/java/com/yourcity/domain/enums/NetworkStatus.java
@@ -1,4 +1,4 @@
-package com.yourcity.yourcity.model.entity.enums;
+package com.yourcity.domain.enums;
 
 public enum NetworkStatus {
     ONLINE,
