+++ domain/src/main/java/com/yourcity/domain/enums/Role.java
@@ -1,0 +1,6 @@
+package com.yourcity.domain.enums;
+
+public enum Role {
+    USER,
+    ADMIN;
+}
