+++ domain/src/main/java/com/yourcity/domain/enums/Type.java
@@ -1,4 +1,4 @@
-package com.yourcity.yourcity.model.entity.enums;
+package com.yourcity.domain.enums;
 
 public enum Type {
     BOULEVARD,
