+++ settings.gradle
@@ -3,4 +3,5 @@ rootProject.name = 'YourCity'
 include 'ui-app'
 include 'your-city-service'
 include 'core'
+include 'domain'
