--- your-city-service/src/main/java/com/yourcity/yourcity/model/entity/Place.java
@@ -1,31 +1,0 @@
-package com.yourcity.yourcity.model.entity;
-
-import com.yourcity.yourcity.model.entity.enums.PlaceCategory;
-import jakarta.persistence.*;
-import lombok.*;
-import lombok.experimental.FieldDefaults;
-import lombok.experimental.FieldNameConstants;
-
-@Data
-@Builder
-@NoArgsConstructor
-@AllArgsConstructor
-@FieldNameConstants
-@FieldDefaults(level = AccessLevel.PRIVATE)
-@Entity
-public class Place {
-
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    Long id;
-
-    String name;
-    String description;
-
-    @Enumerated(EnumType.STRING)
-    PlaceCategory category;
-
-    @ManyToOne(fetch = FetchType.LAZY)
-    @JoinColumn(name = "address_id")
-    Address address;
-}
