--- your-city-service/src/main/java/com/yourcity/yourcity/model/entity/Street.java
@@ -1,23 +1,0 @@
-package com.yourcity.yourcity.model.entity;
-
-import com.yourcity.yourcity.model.entity.enums.Type;
-import jakarta.persistence.*;
-import lombok.*;
-import lombok.experimental.FieldDefaults;
-
-@Data
-@Builder
-@NoArgsConstructor
-@AllArgsConstructor
-@FieldDefaults(level = AccessLevel.PRIVATE)
-@Entity
-public class Street {
-
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    Long id;
-
-    @Enumerated(EnumType.STRING)
-    Type type;
-    String name;
-}
