--- your-city-service/src/main/java/com/yourcity/yourcity/model/entity/enums/Role.java
@@ -1,6 +1,0 @@
-package com.yourcity.yourcity.model.entity.enums;
-
-public enum Role {
-    USER,
-    ADMIN;
-}
