+++ your-city-service/src/main/java/com/yourcity/yourcity/repository/ChatMessageRepository.java
@@ -1,6 +1,6 @@
 package com.yourcity.yourcity.repository;
 
-import com.yourcity.yourcity.model.entity.ChatMessage;
+import com.yourcity.domain.entity.ChatMessage;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
