+++ your-city-service/src/main/java/com/yourcity/yourcity/repository/CityRepository.java
@@ -1,6 +1,6 @@
 package com.yourcity.yourcity.repository;
 
-import com.yourcity.yourcity.model.entity.City;
+import com.yourcity.domain.entity.City;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
