+++ your-city-service/src/main/java/com/yourcity/yourcity/repository/EventRepository.java
@@ -1,6 +1,6 @@
 package com.yourcity.yourcity.repository;
 
-import com.yourcity.yourcity.model.entity.Event;
+import com.yourcity.domain.entity.Event;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.data.jpa.repository.JpaRepository;
